fix: canonicalize iam url (#125)
This change forces canonical use of an IAM URL. The standard operation path for a token
request is now unconditionally added at request time. If a user provides a URL that
includes the path, it will be removed.

As it turns out, there were no tests around URL behavior so I added tests that capture
it, including the behavior added in this change.