feat: support username/apikey use-case in CloudPakForDataAuthenticator (#126)
This commit enhances the CP4D Authenticator to support the username/apikey use case.
With these changes, a user has the ability to construct a CP4D authenticator using
either a username/password pair or a username/apikey pair. The authenticator will
now use the official "POST /v1/authorize" operation for obtaining an access token.

I also added an integration test for interacting with a live service. It is not
executed by default.