fix: stub sdk-test-helpers in browser environments
The "sdk-test-helpers" file contains helper functions that generated SDKs use
in their generated unit tests. These functions are never actually needed when
running standard operational code, therefore they are never used in the browser.

The `expect` library that we use for testing has a dependency that isn't browser
friendly in certain scenarios. This causes the core and any SDKs importing it
to throw errors in these scenarios (like the React framework).

This change stubs the sdk-test-helpers in browser environments the same way we
stub the credentials file reading code.