feat: add support for new ContainerAuthenticator (#151)
This PR adds an authenticator and token manager for the new "container" flavor of authentication. This style retrieves a compute resource token from the file system and uses that to retrieve an IAM access token, which is used to authenticate with the service.

As part of this PR, I refactored the authenticators to match the refactoring in the token manager (moving out IAM specific functionality into a shared class).