feat: add authenticationType method to authenticators
This commit adds a new method ('authenticationType') to
the Authenticator base class, AuthenticatorInterface,
and each authenticator implementation (subclasses of
Authenticator).
The function will return a string indicating the
authenticator type ('basic', 'iam', etc.).
This brings the node core in line with the Go and Jave
cores where the authenticators already support a similar
method.