fix: add new validateParams function
This commit adds the "validateParams" function
to the node core.  This function will be used
by the operation code generated by the Node
SDK generator to validate the parameters found
within a "params" object.  It will check for
required parameters that are not present, and also
check to make sure that all parameters that ARE present
are valid (i.e. the param name is among the expected
valid parameter names for the operation).