fix: avoid Buffer re-encode during compression (#184)
If the data is already a Buffer then just compress it, don't convert it to a string and back to a Buffer again before compressing it.

A test with Buffer input is added. The input is chosen carefully to ensure that if it was assumed to be UTF-8 encoded it would not be the same when decompressed and added an additional assertion for the decompressed data to match the original.