fix: remove 'module' entry from package.json
A recent change to the library introduced
a separate es6 module for the library to support
tree shaking.  Unfortunately, if an SDK project
doesn't (at a minimum) define the 'main' entry point
in its package.json file, then a problem can occur
due to an es5 vs es6 incompatibility.
This commit provides a tactical fix for this issue
by removing the 'module' entry point from the library's
package.json.  This will prevent webpack from being
able to minimize the resulting application code, but
it will avoid the problem.

Once we have a strategy for rolling out "tree
shakability" in various SDK projects, we can
re-introduce the 'module' entry point.