fix: ensure dist/docs is a directory during build
A recent change seems to have removed the creation of the `dist/docs/`
folder during the build and as such, the `cp` command we invoke was
renaming the JSON docs file as a new file with the name `docs`.

This change ensures a folder called `dist/docs/` is created if one
is not already present.