fix: update file-type to resolve vulnerability
The `file-type` package has a vulnerability that persists until v16.5.4. This
commit updates the package to v16.5.4 to avoid the vulnerability. However, the
package update required changes in how we use the package in our code which
resulted in incompatible updates to a couple of functions that are part of our
public API. Though it is unlikely these functions are being widely used, this
change will need to go into a new major version.

BREAKING CHANGE: two synchronous public functions are now asynchronous

The function `getContentType` formerly returned a string but now returns a
Promise that resolves to a string. The function `buildRequestFileObject`
formerly returned a `FileObject` but now returns a Promise that resolves to
a `FileObject`.

Fixes #204

Signed-off-by: Dustin Popp <dpopp07@gmail.com>