fix: correct AuthenticatorInterface type (#206)
The interface incorrectly defined the data channel as "void or Error" when it
should only be "void", as Errors are sent in a different channel. This updates
the type to be correct. It does not constitute a breaking update as we
already defined authenticators with the new type in this project.

Resolves #176

Signed-off-by: Dustin Popp <dpopp07@gmail.com>