fix: allow for synchronous building of file requests
This issue was caused by an oversight in adding an `await` statement within a `forEach`
loop, which will result in asynchronous operations happening without the synchronous
control you'd expect from async/await.

This solves the issue by simply converting the loops to "for of" loops.

Note that the AirBnb style guide we follow prohibits the use of "for of" loops and that
ESLint discourages "await" statements in any kind of loop. However, to adherence to the
rules would require a significant refactor that is not current worth the effort. So for
now, we are ignoring the rules.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>