fix(auth): migrate to secure usage of jwt for token authentication (#225)
There is a vulnerability in v8 of the `jsonwebtoken` dependency. This commit
upgrades to v9 to resolve the vulnerability. Additionally, they made an effort
in this version to discourage the less secure "decode" method in favor of the
more secure "verify" method (1). This commit also refactors the code and tests to
use the "verify" method.

(1) See this PR for context: https://github.com/auth0/node-jsonwebtoken/pull/741

Signed-off-by: Dustin Popp <dpopp07@gmail.com>