fix(auth): revert to using decode instead verify for jwt (#227)
The move to `verify` was a bit shortsighted. It requires a public key or secret as an
argument to verify the signed token. Typical usage of IBM's IAM service doesn't lend
itself to this flow and the tokens returned don't have valid JWT signatures. This led
to a runtime error everytime the core made an IAM request - a bit of a showstopping bug.

Additionally, we never had a goal of performing client-side validation of these tokens,
we only decode them to determine the expiration time for usage in our refresh logic.
The `decode` method is perfectly sufficient for that and indeed is called within the
`verify` method anyways. Perhaps this flow will change in the future but this is all
we need for now.

This reverts the logic to what we were using before but still using the safe version
of the dependency. The new version makes the `decode` function read-only so I had to
adjust our approach to mocking in the unit tests.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>