fix: return json error response body as object
Error response bodies are currently returned as strings, even when the content is JSON.
This is a bug - the SDK should return the result as an object. This commit adds logic
to return the result as an object, converting it if necessary.

It also updates the error object to use the "result" field for the response body, in
order to be consistent with the success path. The "body" field is still set for
backward-compatibility but is marked as deprecated.

Additionally, for both success and error responses, if the content is JSON but the
result cannot be parsed as an object, an exception will be raised, alerting the user
that the service returned a malformatted body.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>