fix: correctly handle empty response bodies
There are cases, such as a HEAD request, where the response body will have
a JSON content type but the body will be empty (i.e. an empty string). This
is technically valid JSON but is not parsable into an object, so the code
throws an exception. This commit adds specific handling for this scenario,
as well as a couple of test cases

Signed-off-by: Dustin Popp <dpopp07@gmail.com>