fix(deps): resolve vulnerability in axios version
This commit forces a release for the updated axios version that
resolved a vulnerability. It also includes build changes to stop
testing on Node 14, which had EOL 6 months ago, start testing on
Node 20, and document that Node 14 is technically no longer
guaranteed, even though it is supported.

This also fixes a test to be more flexible on different Node
versions, as `JSON.parse()` throws slightly different errors
on Node 18 vs Node 20.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>