fix(token-manager): handle request errors when refreshing tokens
This commit introduces a Promise rejection handler to the `getToken`
method where it was previously lacking. The refresh token kick-off did
not catch any exceptions, resulting in the program crashing when a
refresh error occurred. Now, the error will be caught and logged but the
logic will continue as-is and the request will be retried later.

Signed-off-by: Dustin Popp <dpopp07@gmail.com>