fix(deps): remove `expect` as redundant development dependency
We previously exposed the `expect` package as a production-level dependency,
because this package includes unit test utility functions for SDK libraries
that depend on this package. However, `expect` is meant to be a development
dependency and introduces some sizebloat and vulnerability attack surface to
this package.

Fortunately, the `expect` package is intended to be used from `jest` and is
not required to be installed on its own unless specifically needed. Since
all SDK libraries depending on this package also use `jest`, it is redundant
to include the dependency in this package. The test utilities provided will
never be run in production and the SDKs will be able to run their tests
using the already-installed `jest` package. For this reason, `expect` is
removed in this commit.

Signed-off-by: Dustin Popp <dustinpopp@ibm.com>