fix: remove test utilities and dependency from package
The SDK unit test helpers, which relied on test-scoped dependency "expect",
have been removed from the package. They will be exposed through a new NPM
package, separate from this one. The intention is to decouple testing
logic from production logic. This package, which is scoped for production,
will no longer contain the "expect" dependency, which has caused issues for
users. SDK projects will need to install the new package as a development
dependency and update their test code to use it instead. The functionality
will all remain the same.

This new package is still included in this project but it will not be built
and published as a part of public API. Instead, it will be isolated and
self-contained, with its own package.json, etc. For now, it will be published
manually - semantic release will only be used to automate publishing of the core.

BREAKING CHANGE: SDK test helpers are no longer exposed in this package.

To continue using the test helpers, which SDK unit tests still rely on, install
the new "@ibm-cloud/sdk-test-utilities" package as a development dependency.

Signed-off-by: Dustin Popp <dustinpopp@ibm.com>